Add mysql-common to the debian package list
The mysql-common package installs the my.cnf file for MySQL. It is
required that we install this package before installing the my.cnf
file with the template because when this package gets installed,
if it sees a my.cnf pre-existing, it will prompt for options to
either keep the package maintainer version or the user installed
version. Installing this prior to creating the my.cnf file works.